feat: design minimal requried components for mdx
